Added script to load test data
+ Added script to load test data
~ Changed name for image "ssl-enabled-postgres" to "patientenakte-ssl-enabled-postgres"
~ Changed docker-compose to use predefined names for containers
~ Changed build script to produce more detailed output on existing containers

+++ build_docker_images.sh
@@ -1,37 +1,64 @@
 #!/usr/bin/env bash
 
-# build the Docker image for the Authentifizierungsdienst
-docker build -t patientenakte-auth_api ./Authentifizierungsdienst
+echo Building images for project Patientenakte
+echo Note: Any images which are already present, as identified by their docker image tag, will NOT be rebuilt
+echo
+echo
 
-# build the Docker image for the Dateiverwaltungsdienst
-docker build -t patientenakte-file_api ./Dateiverwaltungsdienst
+cd "$(dirname "$0")"
 
-# build the Docker image for the ssl enabled postgres database
-docker build -t ssl-enabled-postgres -f ./Konfiguration/Dockerfiles/ssl-enabled-postgres.txt ./Konfiguration/Dockerfiles
+if [ $(docker image ls | grep "^patientenakte-auth_api " | wc -l) -gt 0 ]; then
+	echo Authentication service image is already present. Skipping build.
+else
+	echo Building authentication service
+	docker build -t patientenakte-auth_api ./Authentifizierungsdienst
+fi
 
-# build the Docker image for the frontend build
-docker build -t patientenakte-build-frontend -f ./Konfiguration/Dockerfiles/frontend-build-environment.txt ./Konfiguration/Dockerfiles
+if [ $(docker image ls | grep "^patientenakte-file_api " | wc -l) -gt 0 ]; then
+	echo File service image is already present. Skipping build.
+else
+	echo Building file service
+	docker build -t patientenakte-file_api ./Dateiverwaltungsdienst
+fi
 
-# start a new Docker container from patientenakte-file_ap
-# docker create -it --name patientenakte-build-frontend ubuntu sh -c "apt-get update && apt-get install npm -y && cd frontend-project && npm install && npm run build"
-docker create -i --name patientenakte-build-frontend patientenakte-build-frontend
+if [ $(docker image ls | grep "^patientenakte-ssl-enabled-postgres " | wc -l) -gt 0 ]; then
+	echo Postgres with TLS image is already present. Skipping build.
+else
+	echo Building Postgres with TLS
+	docker build -t patientenakte-ssl-enabled-postgres -f ./Konfiguration/Dockerfiles/ssl-enabled-postgres.txt ./Konfiguration/Dockerfiles
+fi
 
-# copy the frontend project into the container
-# Cannot exclude from docker cp; need to include explicitly
-docker cp ./Webserver/frontend/public patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/src patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/babel.config.js patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/jsconfig.json patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/package.json patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/package-lock.json patientenakte-build-frontend:/frontend-project/frontend/
-docker cp ./Webserver/frontend/vue.config.js patientenakte-build-frontend:/frontend-project/frontend/
+if [ $(docker image ls | grep "^patientenakte-bff " | wc -l) -gt 0 ]; then
+	echo Webserver image is already present. Skipping build.
+else
+	
+	if [ $(docker image ls | grep "^patientenakte-build-frontend " | wc -l) -gt 0 ]; then
+		echo Frontend build utility image is already present. Skipping build.
+	else
+		echo Building frontend build utility
+		docker build -t patientenakte-build-frontend -f ./Konfiguration/Dockerfiles/frontend-build-environment.txt ./Konfiguration/Dockerfiles
+	fi
+	
+	echo Building webserver
+	
+	docker create -i --name patientenakte-build-frontend patientenakte-build-frontend
+	
+	# Cannot exclude from docker cp; need to include explicitly
+	docker cp ./Webserver/frontend/public patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/src patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/babel.config.js patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/jsconfig.json patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/package.json patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/package-lock.json patientenakte-build-frontend:/frontend-project/frontend/
+	docker cp ./Webserver/frontend/vue.config.js patientenakte-build-frontend:/frontend-project/frontend/
 
-docker start patientenakte-build-frontend
-docker attach patientenakte-build-frontend
-docker stop patientenakte-build-frontend
-# rm -R ./Webserver/static
-docker cp patientenakte-build-frontend:/frontend-project/frontend/dist ./Webserver/static
-docker container rm patientenakte-build-frontend
-
-# build the Docker image for the Webserver
-docker build -t patientenakte-bff ./Webserver
+	docker start patientenakte-build-frontend
+	docker attach patientenakte-build-frontend
+	docker stop patientenakte-build-frontend
+	rm -R ./Webserver/static
+	docker cp patientenakte-build-frontend:/frontend-project/frontend/dist ./Webserver/static
+	docker container rm patientenakte-build-frontend
+	
+	docker build -t patientenakte-bff ./Webserver
+	
+fi
